[DC-2278] Update the cleaning rule to store concept_ids in lookup table
