+++ data_steward/cdr_cleaner/clean_cdr.py
@@ -113,7 +113,7 @@ from cdr_cleaner.cleaning_rules.covid_ehr_vaccine_concept_suppression import Cov
 from cdr_cleaner.cleaning_rules.missing_concept_record_suppression import MissingConceptRecordSuppression
 from cdr_cleaner.cleaning_rules.create_deid_questionnaire_response_map import CreateDeidQuestionnaireResponseMap
 from cdr_cleaner.cleaning_rules.vehicular_accident_concept_suppression import VehicularAccidentConceptSuppression
-from cdr_cleaner.cleaning_rules.deid.missing_and_replaced_concept_suppression import \
+from cdr_cleaner.cleaning_rules.deid.ct_replaced_concept_suppression import \
     ControlledTierReplacedConceptSuppression
 from constants.cdr_cleaner import clean_cdr_engine as ce_consts
 from constants.cdr_cleaner.clean_cdr import DataStage
